03.09.23 21:02

+++ src/main/java/com/clevertecbank/entity/Transaction.java
@@ -16,7 +16,7 @@ public class Transaction {
     private long id;
     private long sender_id;
     private long receiver_id;
-    private int transaction_value;
+    private double transaction_value;
     private Timestamp date;
     private Transaction_type type;
 }
